requirements.txt
nextcord~=2.0.0a6
PyYAML~=6.0
litematica-tools~=1.0.1